Get GitHub Repositories function is about retrieving a list of repositories, which is typically considered a broader repository-related
