+++ app/controllers/github_controller.py
@@ -10,19 +10,6 @@ from app.controllers.auth_controller import get_installation_access_token
 
 github_controller = Blueprint('github', __name__)
 
-def get_github_repositories(username, access_token):
-    url = f'https://api.github.com/users/{username}/repos'
-    headers = {'Authorization': f'Bearer {access_token}'}
-
-    try:
-        response = requests.get(url, headers=headers, timeout=60)
-        response.raise_for_status()
-        repositories = response.json()
-        return repositories
-    except RequestException as e:
-        logging.error(f"Failed to fetch repositories: {e}")
-        return None
-
 def get_branches(owner, repo_name):
     """Fetches all branches of a repository."""
     access_token = get_installation_access_token()
